+++ src/check-npm.js
@@ -15,7 +15,9 @@ if (Meteor.isClient) {
       'body-parser': '^1.15.2',
       express: '^4.14.0',
       graphql: '^0.9.0',
+      'graphql-subscriptions': '^0.3.0',
       'graphql-tools': '^0.10.0',
+      'subscriptions-transport-ws': '^0.5.5-alpha.0',
     },
     'apollo'
   );