+++ CHANGELOG.md
@@ -3,13 +3,14 @@ All notable changes to this project will be documented in this file. [*File synt
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## vNEXT
-### Added
+### Changed
 - Move logic code to `/src`.
 - Use `rawCollection` method on `Meteor.users` for the current user lookup, to allow the tests to pass (also no need to wrap the GraphQL server in `Meteor.bindEnvironment` as it no more a Fibers-aware code). 
 
 ### Added
 - Set up ESLint with prettier & Meteor settings.
 - Provide `handleDone` utility for asynchronous tests which to catch errors.
+- Handle GraphQL subscriptions client-side with the option `enableSubscriptions` on `createMeteorNetworkInterface` & start a websocket server if a `pubsub` mechanism is passed along the `schema` to `createApolloServer`.
 
 ## [0.5.0] - 2017-03-03
 ### Changed